Add files via upload
Question:
A leap year is a calendar year containing an additional day added to keep the calendar year synchronized with the astronomical or seasonal year. In the gregorian Feb to 29 days rather than common 28. These extra days occur in years that are multiples of four (with the exception of centennial years not divisible by 400). What is the program asking for a year and calculates if this year is a leap year or not?