Add files via upload
Body mass index (BMI) is a value derived from the mass (weight) and height of a person. the BMI is defined as the boidy mass divided by the square of the body height and is universally expressed in units of kg/m2. What is the program which is aska for the length and weight of a person and returns the BMI with the category?